A dictionary that maps raw data column names and values to prettified strings for UI display

Used in conjunction with propalloc::apply_labels() to apply the dictionary's mapping to a data set

dictionary

Format

A data frame with 1727 rows and 6 columns:

dataset

String: Identifier of the dataset. Function apply_labels() takes as arguments the data.frame of the raw data to be prettified, as well as the dataset string (amongst other arguments) to which it is assigned. Therefore, the dataset string need not match exactly the symbol of the data.frame of the raw data, although this might be the most-consistent approach. For example, this would be a valid call: apply_labels(data = bu_rels, dict = propalloc::dictionary, dataset_name = "business_unit_rels") assuming that the dataset value "business_unit_rels" appears in propalloc::dictionary. The most-consistent approach would be to use "bu_rels" as the dataset identifier, as in: apply_labels(data = bu_rels, dict = propalloc::dictionary, dataset_name = "bu_rels")

variable

String: Field name of the dataset which is to be prettified. eg. "bu".

variable_label

String: The prettified label to replace variable with for UI display. eg. "Business Unit".

value

String: If NA, then this field is not used. Else, the dictionary will also attempt to prettify values within the variable field. String values in this field should complete all possible values that the variable can take. eg. "bu_a", "bu_b".

value_label

String: The prettified label to replace value with for UI display. eg. "Business Unit A", "Business Unit B". If NA, then this field is not used. If value field is NA, then this field should be too.

value_order

Numeric: An integer value used to order the prettified data.frame according to the appearance of values in field value.