Header Buttons Server Module

Header Buttons Server Module

header_buttons(
  input,
  output,
  session,
  parent_session = NULL,
  include_tour = TRUE,
  tour_steps = NULL,
  include_refresh = TRUE,
  include_help = TRUE,
  include_disclaimer = TRUE,
  help_path = fs::path_package("propalloc", "reports/help.Rmd"),
  disclaimer_path = fs::path_package("propalloc", "reports/disclaimer.Rmd")
)

Arguments

input

shiny input

output

shiny output

session

shiny session

parent_session

shiny session of the parent environment where this module is called from. Only used if include_tour is TRUE.

include_tour

logical - include a 'tour app' button?

tour_steps

list of 'steps' for tour. Passed to rintrojs::introjs().

include_refresh

logical - include a 'refresh' button?

include_help

logical - include a 'help' button?

include_disclaimer

logical - include a 'disclaimer' button?

help_path

character - if include_help is TRUE; specify the path to the 'help' R Markdown document to utilize when button is pressed.

disclaimer_path

character - if include_disclaimer is TRUE; specify the path to the 'disclaimer' R Markdown document to utilize when button is pressed.

Value

server