Reverse Labels

Reverses link[apply_labels()].

reverse_labels(
  data,
  dict,
  from = "value",
  to = "value_label",
  by = "variable",
  names_from = "variable",
  names_to = "variable_label",
  dataset_name = NULL,
  ...
)

Arguments

data
dataset to apply labeling on.

dict
dictionary to use for application of labeling.

from
a column name or position defining words or keys to be replaced

to
a column name or position defining replacement values

by
character or integer - which column in dict defines the columns in data corresponding to each section of the dict. Defaults to "variable".

names_from
column name or position defining where to match names from. Defaults to "variable".

names_to
column name or position defining replacements for names(data). Defaults to "variable_label".

dataset_name
(optional) name of dataset to filter dict by

...
Passed to match_df

Value

un-labeled, R-friendly tibble